Splint webpack config into two (server/client), make config.js return a function which expects a target
