+++ docs/universal-react.md
@@ -1,0 +1,10 @@
+# Universal React App
+
+This template will set up a React project that renders on both the client and server with Node.js.
+It comes with React-Router by default.
+
+Some of the configuration options have different default values. The client entry point is now in _/app/client.js_
+
+There are some additional configuration options:
+
+-   `serverEntryPoint` (_/app/server.js_) - Entry point for your server.
