+++ scripts/webpack/webpack.config.client.js
@@ -1,0 +1,3 @@
+const getWebpackConfig = require('./getWebpackConfig');
+
+module.exports = getWebpackConfig('web');
