+++ scripts/webpack/webpack.config.server.js
@@ -1,0 +1,3 @@
+throw new Error('Not implemented');
+// const getWebpackConfig = require('./getWebpackConfig');
+// module.exports = getWebpackConfig('node');
